+++ src/components/common/OutlinedButton.tsx
@@ -7,6 +7,7 @@ interface OutlinedButtonProps {
   className?: string;
   children?: React.ReactNode;
   ref?: Ref<HTMLButtonElement>;
+  type?: string;
 }
 
 export const OutlinedButton = ({
@@ -16,6 +17,7 @@ export const OutlinedButton = ({
   className,
   children,
   ref,
+  type = "button",
 }: OutlinedButtonProps) => {
   const buttonClassName = ` flex rounded-md justify-center items-center gap-2 w-full h-full p-2 !outline-0 border border-mainBorder dark:border-mainBorderDark bg-outlinedButtonBg dark:bg-outlinedButtonBgDark hover:bg-outlinedButtonBgHover dark:hover:bg-outlinedButtonBgHoverDark text-primaryText dark:text-primaryTextDark dark:stroke-grayIconDark dark:fill-grayIconDark fill-grayIcon stroke-grayIcon ${className}`;