+++ src/components/views/calendar/types.ts
@@ -1,10 +1,24 @@
+import { EventDropArg, EventApi } from "@fullcalendar/core";
+
 export type CalendarEvent = {
   id: string;
   title: string;
-  start: string;
-  end?: string;
+  start: string | undefined;
+  end?: string | undefined; 
 };
+export interface FullCalendarEvent extends CalendarEvent {
+  remove: () => void;
+}
 
 export interface CalendarViewProps {
   calendarEvents: CalendarEvent[];
+}
+
+
+export type CalendarAction = "delete" | "move" | null
+
+export interface EventChange {
+  oldEvent: EventApi;
+  newStart: Date;
+  newEnd: Date | undefined;
 }