+++ src/layout/navbar/Navbar.tsx
@@ -25,7 +25,7 @@ import { useSession } from "../../hooks/auth/useSession";
 import { useHandleLogout } from "../../hooks/auth/useHandleLogout";
 import { useTooltip } from "../../hooks/useTooltip";
 import { Tooltip } from "../../components/common/Tooltip";
-import { LogoutModal } from "./LogoutModal";
+import { LogoutModal } from "../../components/auth/LogoutModal";
 
 export const Navbar = () => {
   const [isLoginModalOpen, setIsLoginModalOpen] = useState(false);